ndt_list: add some convenience functionality.

+++ src/common/list.h
@@ -25,13 +25,13 @@
 
 typedef struct ndt_list ndt_list;
 
-ndt_list* ndt_list_init  (                                            );
-size_t    ndt_list_count (const ndt_list *list                        );
-void*     ndt_list_item  (const ndt_list *list,             size_t idx);
-void      ndt_list_insert(      ndt_list *list, void *item, size_t idx);
-void      ndt_list_add   (      ndt_list *list, void *item            );
-void      ndt_list_rem   (      ndt_list *list, void *item            );
-void      ndt_list_close (      ndt_list **ptr                        );
+ndt_list* ndt_list_init  (                                         );
+size_t    ndt_list_count (const ndt_list *list                     );
+void*     ndt_list_item  (const ndt_list *list,             int idx);
+void      ndt_list_insert(      ndt_list *list, void *item, int idx);
+void      ndt_list_add   (      ndt_list *list, void *item         );
+void      ndt_list_rem   (      ndt_list *list, void *item         );
+void      ndt_list_close (      ndt_list **ptr                     );
 void      ndt_list_sort  (      ndt_list *list,
                           size_t w, int (*c)(const void*, const void*));
